+++ packages/corbado_auth/lib/src/blocks/completed_block.dart
@@ -1,7 +1,5 @@
 import 'package:corbado_auth/src/blocks/types.dart';
 import 'package:corbado_auth/src/process_handler.dart';
-import 'package:corbado_auth/src/services/corbado/corbado.dart';
-import 'package:corbado_auth/src/types/screen_names.dart';
 import 'package:corbado_frontend_api_client/corbado_frontend_api_client.dart';
 
 class CompletedBlockData {