+++ packages/passkeys/passkeys_ios/pigeons/messages.dart
@@ -87,6 +87,8 @@ class AuthenticateResponse {
 abstract class PasskeysApi {
   bool canAuthenticate();
 
+  bool hasBiometrics();
+
   @async
   RegisterResponse register(
     String challenge,