+++ packages/passkeys/passkeys_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.5.0
+* Added `getAvailability` api to check for passkey support
+* Upgrade `androidx.credentials:credentials` to 1.3.0
+
 ## 2.4.0
 * Added handling for `CreateCredentialNoCreateOptionException` when no viable creation options were found